Create reference to a GitHub repository

This function constructs a list of needed information to send API calls to a specific GitHub repository. Specifically, it stores information on thhe repository's name and owner, the type (whether or not Enterprise GitHub), and potentially credentials to authenticate.

create_repo_ref(repo_owner, repo_name, is_enterprise = FALSE,
  hostname = "", identifier = "", password = "")

Arguments

repo_owner
Repository owner's eid or GitHub Organization name

repo_name
Repository name

is_enterprise
Boolean denoting whether or not working with Enterprise GitHub.Defaults to FALSE

hostname
Host URL stub for Enterprise repositories (e.g. if URL is "github.abc.com" then set to "abc")

identifier
GitHub username. Ideally should be left blank and then defaults to GITHUB_PAT or GITHUB_ENT_PAT environment variables if not provided.

password
GitHub password. Ideally should be left blank and then defaults to empty string which works in conjunction with personal access tokens. Required if supplying username.

Value

List of repository reference information and credentials

Examples

# NOT RUN {
tidytracker <- create_repo_ref('emilyriederer', 'tidytracker')
# }