Save SVG of Agile-style task board of issue status with links to issues

This function creates the same plot as viz_taskboard then edits the underlying XML so that the "cards" are linked to the corresponding issues on GitHub. It saves a file with the reuslting SVG, which can then be read into an RMarkdown HTML document as shown in the Examples.

viz_taskboard_links(g, filepath)

Arguments

g
ggplot2 object returned by viz_gantt_closed()

filepath
Location to save resulting SVG file of ggplot2, if desired. Leave blank for function to output message precisely as needed to render in HTML RMarkdown with chunk option results = 'asis'

Value

SVG version of ggplot2 object with links to relevant GitHub issues. Either writes output to file or to console (to be captured in RMarkdown) depending on existence of filepath argument

Details

Credit goes to this Stack Overflow answer for figuring out how to do this: https://stackoverflow.com/questions/42259826/hyperlinking-text-in-a-ggplot2-visualization/42262407

See also

Examples

# NOT RUN {
# In R, to save to file:
viz_taskboard_links(issues, "my_folder/my_file.svg")

# In RMarkdown chunk, to print as output:
```{r results = 'asis', echo = FALSE}
g <- viz_taskboard(issues)
viz_taskboard_links(g)
````
# }